requirements.txt
astroid==2.3.2
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mdx-linkify==1.2
Pillow==6.2.1
pkg-resources==0.0.0
pylint==2.4.3
python-slugify==4.0.0
PyYAML==5.1.2
requests==2.22.0
requests-cache==0.5.2
six==1.12.0
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.6
webencodings==0.5.1
wrapt==1.11.2
xmltodict==0.12.0
